Добавлена верификация через Email
