+++ app/Http/Controllers/Auth/RegisterController.php
@@ -67,6 +67,7 @@ class RegisterController extends Controller
             'name' => $data['name'],
             'email' => $data['email'],
             'password' => Hash::make($data['password']),
+            'role' => User::ROLE_READER,
         ]);
     }
 }
